feat(templates): Removed error-inducing text
The text for setting up an appointment may have been misleading in the
sense that customers may actually show up at the location without having
set an appointment. It is now replaced by a simple prompt requesting the
desired day and time from the user. A human agent with access to the
whatsapp number from a cellphone will further the interaction.

also some minor renaming and formatting stuff.